+++ src/ui/common/editors/Needle.h
@@ -14,3 +14,6 @@ bool Editor(const char* label, hh::needle::WorldRenderingPipeline& pipeline);
 bool Editor(const char* label, hh::needle::RenderingPipelineRangers& pipeline);
 bool Editor(const char* label, hh::needle::RenderUnit& renderUnit);
 bool Editor(const char* label, hh::needle::RenderTextureHandle& renderTexture);
+bool Editor(const char* label, hh::needle::MeshResource& meshResource);
+bool Editor(const char* label, hh::needle::InstanceParameterContainerData& ipcd);
+bool Editor(const char* label, hh::needle::ParameterValueObjectContainer& pvo);